Sandwiches altered - price to bigdecimal

+++ src/main/java/be/ucll/da/dentravak/model/Application.java
@@ -8,10 +8,11 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
+import java.math.BigDecimal;
+
 @SpringBootApplication
 public class Application {
-
-
+    
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
     }
@@ -19,11 +20,8 @@ public class Application {
     @Bean
     public CommandLineRunner commandLineRunner(@Autowired LunchRepository lunchRepository) {
         return args -> {
-
-            lunchRepository.save(new Sandwich("Broodje gezond", 3.5, "Brood, kaas, hesp, ei, sla"));
-            lunchRepository.save(new Sandwich("Smos", 2.8, "Brood, boulet, ei, sla, mayonnaise"));
-
-
+            lunchRepository.save(new Sandwich("Broodje gezond", BigDecimal.valueOf(3.5), "Brood, kaas, hesp, ei, sla"));
+            lunchRepository.save(new Sandwich("Smos", BigDecimal.valueOf(2.8), "Brood, boulet, ei, sla, mayonnaise"));
         };
     }
 }
